energy sword fixes
- It was previously possible to combine two dualsabers or a dualsaber and an esword to make a dualsaber (ik, it doesn't make sense). It's fixed now.
- Dualsaber can only be activated when it is dual-wielded, and will deactivate if the user picks up an item in their other hand.

+++ code/game/objects/items/weapons/melee/energy.dm
@@ -41,8 +41,8 @@
 /obj/item/weapon/melee/energy/attack_self(mob/living/user as mob)
 	if(active)
 		if((MUTATION_CLUMSY in user.mutations) && prob(50))
-			user.visible_message("<span class='danger'>\The [user] accidentally cuts \himself with \the [src].</span>", \
-			"<span class='danger'>You accidentally cut yourself with \the [src].</span>")
+			user.visible_message(SPAN("danger","\The [user] accidentally cuts \himself with \the [src]."), \
+			SPAN("danger","You accidentally cut yourself with \the [src]."))
 			user.take_organ_damage(5, 5)
 		deactivate(user)
 	else
@@ -98,12 +98,12 @@
 /obj/item/weapon/melee/energy/axe/activate(mob/living/user)
 	..()
 	icon_state = "axe1"
-	to_chat(user, "<span class='notice'>\The [src] is now energised.</span>")
+	to_chat(user, SPAN("notice","\The [src] is now energised."))
 
 /obj/item/weapon/melee/energy/axe/deactivate(mob/living/user)
 	..()
 	icon_state = initial(icon_state)
-	to_chat(user, "<span class='notice'>\The [src] is de-energised. It's just a regular axe now.</span>")
+	to_chat(user, SPAN("notice","\The [src] is de-energised. It's just a regular axe now."))
 
 /*
  * Energy Sword
@@ -149,7 +149,7 @@
 
 /obj/item/weapon/melee/energy/sword/activate(mob/living/user)
 	if(!active)
-		to_chat(user, "<span class='notice'>\The [src] is now energised.</span>")
+		to_chat(user, SPAN("notice","\The [src] is now energised."))
 		mod_shield = 2.5
 	..()
 	attack_verb = list("attacked", "slashed", "stabbed", "sliced", "torn", "ripped", "diced", "cut")
@@ -157,7 +157,7 @@
 
 /obj/item/weapon/melee/energy/sword/deactivate(mob/living/user)
 	if(active)
-		to_chat(user, "<span class='notice'>\The [src] deactivates!</span>")
+		to_chat(user, SPAN("notice","\The [src] deactivates!"))
 		mod_shield = 1.0
 	..()
 	attack_verb = list()
@@ -165,6 +165,8 @@
 
 /obj/item/weapon/melee/energy/sword/attackby(obj/item/sword as obj, mob/user as mob)
 	if(istype(sword, /obj/item/weapon/melee/energy/sword))
+		if (istype(sword,/obj/item/weapon/melee/energy/sword/dualsaber) || istype(src,/obj/item/weapon/melee/energy/sword/dualsaber))
+			return
 		to_chat(user, "<span class='notice'>You attach the ends of the two energy swords, making a single double-bladed weapon!</span>")
 		new /obj/item/weapon/melee/energy/sword/dualsaber(user.loc)
 		qdel(sword)
@@ -230,8 +232,20 @@
 	blade_color = "purple"
 
 /obj/item/weapon/melee/energy/sword/dualsaber/activate(mob/living/user)
+	if(!is_held_twohanded(user))
+		to_chat(user, SPAN("warning","You must hold \the [src] with both hands to activate it."))
+		return
 	..()
 	icon_state = "dualsaber[blade_color]"
+
+/obj/item/weapon/melee/energy/sword/dualsaber/update_twohanding()
+	if(active)
+		if (!isliving(loc))
+			return ..()
+		var/mob/living/user = loc
+		if (!is_held_twohanded(user))
+			deactivate(user)
+	..()
 
 /*
  *Energy Blade
